requirements.txt
fastapi==0.111.1
langchain==0.2.11
langchain_community==0.2.10
langchain_openai==0.1.19
langchain_text_splitters==0.2.2
openai==1.37.1
pydantic==1.10.2
PyPDF2==3.0.1
python-dotenv==1.0.1
typing_extensions==4.12.2
